Crate cdm_rt

Startup code and minimal runtime for the CDM-16 processor.

Features

This crate provides:

Default interrupt and exception handlers using the halt instruction

A macro for defining the entry point of the program: entry

A macro for redefining exception handlers: exception

A macro for defining interrupt handlers: interrupt

A macro for registering interrupt handlers in the IVT: interrupt_vectors

Cargo features

interrupts

This feature disables the default interrupt handlers and allows the use of interrupt_vectors

harvard

This feature enables generation of images for the Harvard architecture with separate program and data address spaces. By default, the von Neumann architecture with one address space is used.

Requirements

Rust flags

The crate generates a linker script named link.x in the output directory. It needs to be passed to the linker. This can be done by passing -Clink-arg=-Tlink.x to rustc.

Entry point

Exactly one function needs to be marked as the application entry point by applying the entry attribute.

Example

.cargo/config.toml

[build]
target = "cdm-none"
rustflags = [ "-Clink-arg=-Tlink.x" ]
[unstable]
build-std = [ "core" ]

src/main.rs

#![no_std]
#![no_main]

use cdm_rt::entry;

#[entry]
fn main() -> ! {
    loop { /* .. */ }
}

Memory mapped I/O

If you want to use memory mapped I/O, you will need to:

create a linker script (e.g. memory.x in the project root)

set the value of RAM_ORIGIN and/or RAM_LENGTH in the linker script to reserve a range of addresses for the MMIO registers

create one or more symbols for the registers and set their values in the linker script

pass the script to the linker (e.g. pass -Clink-arg=-Tmemory.x to rustc)

declare the symbols in Rust inside an extern "C" block

use core::ptr::read_volatile and/or core::ptr::write_volatile to read from and write to the registers

Important note: when using the von Neumann architecture, RAM_ORIGIN and MMIO register addresses need to be greater or equal to 0x100 (256 in decimal) to avoid overlapping with the IVT.

Example

.cargo/config.toml

[build]
target = "cdm-none"
rustflags = [ "-Clink-arg=-Tlink.x", "-Clink-arg=-Tmemory.x" ]
[unstable]
build-std = [ "core" ]

memory.x

RAM_ORIGIN = 0x120;

MMIO_IN = 0x100;
MMIO_OUT = 0x102;

src/mmio.rs

unsafe extern "C" {
    static MMIO_IN: u16;
    static mut MMIO_OUT: u16;
}

pub fn get_in() -> u16 {
    unsafe { core::ptr::read_volatile(&raw MMIO_IN) }
}

pub fn set_out(value: u16) {
    unsafe { core::ptr::write_volatile(&raw mut MMIO_OUT, value) }
}

Macros

interrupt_vectorsinterrupts

Defines the application-specific interrupt handler section of the interrupt vector table.

Structs

InterruptVector

Represents a vector in the interrupt vector table.

Psr

Processor status register bit field.

Constants

EXCEPTION_COUNT

The number of exception vectors in the interrupt vector table, including the start vector.

INTERRUPT_COUNT

The number of application-specific interrupt vectors in the interrupt vector table.

Attribute Macros

entry

Defines the entry point of the program.

exception

Defines an exception handler.

interrupt

Defines an application-specific interrupt handler.