cdm_rt

Macro interrupt_vectors

macro_rules! interrupt_vectors {
    ($($elems:expr),* $(,)?) => { ... };
}

Available on crate feature interrupts only.

Defines the application-specific interrupt handler section of the interrupt vector table.

The interrupt vectors specified in the arguments are placed sequentially after the reset and exception vectors, starting from index INTERRUPT_START. The rest of the table is filled with InterruptVector::DEFAULT.

Must be used once in the dependency graph.

Use the interrupt attribute to define interrupt handler functions.

interrupt_vectors![
    InterruptVector(MyHandler1, Psr::None), // int INTERRUPT_START+0
    InterruptVector(MyHandler2, Psr::None), // int INTERRUPT_START+1
    InterruptVector(MyHandler3, Psr::None), // int INTERRUPT_START+2
];

#[cdm_rt::interrupt]
fn MyHandler1() { /* .. */ }
#[cdm_rt::interrupt]
fn MyHandler2() { /* .. */ }
#[cdm_rt::interrupt]
fn MyHandler3() { /* .. */ }