cdm_rt

Struct InterruptContext

pub struct InterruptContext(/* private fields */);

Represents the processor context saved on the stack when an ISR is entered.

Implementations

impl InterruptContext

pub unsafe fn as_mut(&mut self) -> &mut CpuContext

Gives mutable access to the contents of the interrupt context.

Safety

This function is unsafe since modifying the content of the context can easily lead to undefined behavior. For example, by writing an invalid value to the program counter field, the CPU can jump to arbitrary code at the end of the interrupt.