cdm_rt

Struct Psr

#[repr(transparent)]
pub struct Psr(pub u16);

Processor status register flags.

Tuple Fields
0: u16

The bit representation of the processor status value.

Implementations

impl Psr

pub const NONE: Psr

No flags.

pub const ARITH_NEGATIVE: Psr

Arithmetic negative flag.

pub const ARITH_ZERO: Psr

Arithmetic zero flag.

pub const ARITH_OVERFLOW: Psr

Arithmetic overflow flag.

pub const ARITH_CARRY: Psr

Arithmetic carry flag.

pub const CONTEXT_BITS: Psr

Bit mask for the MMU context number (8 consecutive bits).

pub const IO_ENABLE: Psr

I/O mapping enable flag.

pub const MODE_USER: Psr

Processor mode flag (system = 0, user = 1).

pub const INTERRUPT_ENABLE: Psr

Interrupt enable flag.

impl Psr

pub const fn from_bits(bits: u16) -> Psr

Creates a processor status value from the corresponding bit representation.

pub const fn bits(self) -> u16

Gets the bit representation of the processor status value.

pub const fn contains(self, flags: Psr) -> bool

Checks if the processor status value has all of the specified flags.

pub const fn intersects(self, flags: Psr) -> bool

Checks if the processor status value has any of the specified flags.

pub const fn not(self) -> Psr

Inverts the flags of the processor status value.

pub const fn and(self, other: Psr) -> Psr

ANDs the flags of two processor status values.

pub const fn or(self, other: Psr) -> Psr

ORs the flags of two processor status values.

pub const fn xor(self, other: Psr) -> Psr

XORs the flags of two processor status values.

pub const fn with_context_number(self, context_number: u8) -> Psr

Replaces the MMU context number in a processor status value.

pub const fn from_context_number(context_number: u8) -> Psr

Creates a processor status value from a MMU context number.

pub const fn context_number(self) -> u8

Extracts the MMU context number from the processor status value.