cdm_rt

Struct Psr

pub struct Psr { /* private fields */ }

Processor status register flags.

Implementations

impl Psr

pub const None: Psr

No flags.

pub const ArithNegative: Psr

Arithmetic negative flag.

pub const ArithZero: Psr

Arithmetic zero flag.

pub const ArithOverflow: Psr

Arithmetic overflow flag.

pub const ArithCarry: Psr

Arithmetic carry flag.

pub const Interrupt: Psr

Interrupt enable flag.

pub const fn bits(&self) -> u16

Returns the underlying bits of the bitmask.

pub const fn all_bits() -> Psr

Returns a bitmask that contains all values.

This will include bits that do not have any flags. Use ::all_flags() if you only want to use flags.

pub const fn all_flags() -> Psr

Returns a bitmask that contains all flags.

pub const fn is_all_bits(&self) -> bool

Returns true if the bitmask contains all values.

This will check for bits == !0, use .is_all_flags() if you only want to check for all flags

pub const fn is_all_flags(&self) -> bool

Returns true if the bitmask contains all flags.

This will fail if any unused bit is set, consider using .truncate() first.

pub const fn all() -> Psr

👎Deprecated: Please use the ::all_bits() constructor

Returns a bitmask that contains all values.

This will include bits that do not have any flags. Use ::all_flags() if you only want to use flags.

pub const fn is_all(&self) -> bool

👎Deprecated: Please use the .is_all_bits() method

Returns true if the bitmask contains all values.

This will check for bits == !0, use .is_all_flags() if you only want to check for all flags

pub const fn full() -> Psr

👎Deprecated: Please use the ::all_flags() constructor

Returns a bitmask that contains all flags.

pub const fn is_full(&self) -> bool

👎Deprecated: Please use the .is_all_flags() method

Returns true if the bitmask contains all flags.

This will fail if any unused bit is set, consider using .truncate() first.

pub const fn none() -> Psr

Returns a bitmask that does not contain any values.

pub const fn is_none(&self) -> bool

Returns true if the bitmask does not contain any values.

pub const fn truncate(&self) -> Psr

Returns a bitmask that only has bits corresponding to flags

pub const fn intersects(&self, other: Psr) -> bool

Returns true if self intersects with any value in other, or if other does not contain any values.

This is equivalent to (self & other) != 0 || other == 0.

pub const fn contains(&self, other: Psr) -> bool

Returns true if self contains all values of other.

This is equivalent to (self & other) == other.

pub const fn not(self) -> Psr

Returns the bitwise NOT of the bitmask.

pub const fn and(self, other: Psr) -> Psr

Returns the bitwise AND of the bitmask.

pub const fn or(self, other: Psr) -> Psr

Returns the bitwise OR of the bitmask.

pub const fn xor(self, other: Psr) -> Psr

Returns the bitwise XOR of the bitmask.